% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MerfishData-package.R
\docType{package}
\name{MerfishData-package}
\alias{MerfishData-package}
\alias{MerfishData}
\title{Collection of public MERFISH datasets}
\description{
MerfishData is an ExperimentHub package that serves publicly available
datasets obtained with Multiplexed Error-Robust Fluorescence in situ Hybridization
(MERFISH). MERFISH is a massively multiplexed single-molecule imaging technology
capable of simultaneously measuring the copy number and spatial distribution
of hundreds to tens of thousands of RNA species in individual cells. The scope
of the package is to provide MERFISH data for benchmarking and analysis.
}
\author{
Ludwig Geistlinger \href{mailto:ludwig_geistlinger@hms.harvard.edu}{ludwig_geistlinger@hms.harvard.edu}
}
