% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotRasterImage}
\alias{plotRasterImage}
\title{Plot raster image}
\usage{
plotRasterImage(img)
}
\arguments{
\item{img}{a raster object representing a bitmap image.}
}
\value{
A ggplot object.
}
\description{
Small helper function to plot a raster image.
}
\examples{
 hgrid <- hcl(0, 80, seq(50, 80, 10))
 img <- as.raster(matrix(hgrid, nrow = 4, ncol = 5))
 plotRasterImage(img)
}
